--- fandomk/src/App.module.css
@@ -1,7 +1,0 @@
-.title {
-  color: white;
-}
-
-.text {
-  color: white;
-}
